HBASE-26553. Move JWKS loader to upper level, fix dependencies
